Get the Scene Editor Plugin Working
This commit adds the Scene Editor Plugin. This plugin is the core to
to the functionality of the Mystic Tower game engine. Instead of coding
scenes the Scene Editor Plugin will allow the user to create their scene in
the Phaser (3.52.0) game engine and write that scene to a file that can
then be loaded by the Scene Editor Plugin. To run the game the Scene
Editor plugin will be run in game mode and load the starting scene from
its scene file. This idea might be extended to make the game similar to
a to playable game engine.