Implement All of a Sparse Set class
This commit implements all necessary functionality for the Sparse Set
implementation. Unit tests for the Sparse set were added and can be run
with `npm run test`. This data structure can be used to represent
entities and the associated components can be tracked with an array
lined up and managed in accordance with the dense array. This should be
added as the next step toward implementing an entity component system.