Create Entity Component System (ECS)
This commit uses the sparse set data structure with a modification to
the API to make an Entity Component System (ECS). This will be used by
the rest of the game to make the coding experience of the game easier.
Also the name of the Plugin that is used to run and edit the game was
renamed to the MysticPlugin and the blank starting scene was changed in
turn to the MysticScene.